+++ js/modules/handleLogin.js
@@ -1,3 +1,5 @@
+import { indicator } from "./renderView.js";
+
 const LoginForm = document.getElementById("login-form");
 
 LoginForm.addEventListener("submit", (e) => {
@@ -24,13 +26,7 @@ function loginAuth(database, user, form) {
             window.location.href = '../pages/main.html';
             return
     } else {
-        const container = document.createElement('div');
-        container.innerHTML = `<span class='invalid-span' >Invalid Email or password</span>`
-        form.prepend(container);
-
-        setTimeout(() => {
-            form.removeChild(container);
-        }, 1000);
+      indicator(form, 'Invalid Email or password')
     }
   });
 }
